pyproject.toml
[tool.poetry]
name = "arxiv-gpt"
version = "0.1.0"
description = ""
authors = ["shtmjp <shtm.1058@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
arxiv = "^2.1.3"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
discord-webhook = "^1.3.1"
gspread = "^6.1.4"
oauth2client = "^4.1.3"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
